Some updates to styling as per Dean's suggestions.
